feat(client): show related fields in field form
