+++ packages/amplication-client/src/Entity/SchemaField.tsx
@@ -3,7 +3,7 @@ import { capitalCase } from "capital-case";
 import { ToggleField, TextField } from "@amplication/design-system";
 import EntitySelectField from "../Components/EntitySelectField";
 import EnumSelectField from "../Components/EnumSelectField";
-import RelatedEntityFieldField from "../Components/RelatedEntityFieldField";
+import RelatedEntityFieldField from "./RelatedEntityFieldField";
 import RelationAllowMultipleField from "../Components/RelationAllowMultipleField";
 import { Schema } from "@amplication/data";
 import OptionSet from "../Entity/OptionSet";
@@ -92,7 +92,9 @@ export const SchemaField = ({
           );
         }
         case "#/definitions/EntityFieldId": {
-          return <RelatedEntityFieldField />;
+          return (
+            <RelatedEntityFieldField entityDisplayName={entityDisplayName} />
+          );
         }
         case "#/definitions/RelationAllowMultiple": {
           return (
